Green do_not_delete_period_pricing_pax_when_code_is_empty

+++ src/Domain/UseCases/PeriodPricingPax/DeletePeriodPricingPax.cs
@@ -7,6 +7,9 @@ public class DeletePeriodPricingPax{
     }
 
     public async Task Execute(string Code) {
+        if (string.IsNullOrEmpty(Code)) {
+            throw new ArgumentException("Code is required");
+        }
         await availabilitySynchronizerApiClient.DeletePeriodPricingPax(Code);
     }
 }
